bp: Use dynamic port ranges for ExternalTestCluster
https://github.com/elastic/elasticsearch/commit/5ba4f5fb3c9
